Externalize path
The patch was included in the LaTeX document. I just wanted to use
`patch` to apply the patch to my local installation. This meant copying
it out of the LaTeX document instead of downloading it right from
GitHub.

I have externalized the patch into an external file and included that
into the document using `\inputminted`. It still compiles on my machine.