chore(ci): switch to l3build
l3build is designed to help build and test packages. This commit adds
the `build.lua` configuration file and otherwise only reorganises the
files in such a way as to allow them to be built by l3build.

The CI workflow has also been updated to make use of
`teatimeguest/setup-textlive-action` which allows for the tests to be
run on all operating systems.

Signed-off-by: JP-Ellis <josh@jpellis.me>